feat(frontend): route events UI through live api client
